rename root directory
