+++ mancala-tui/justfile
