+++ mancala-tui/src/board.txt
