+++ mancala-tui/src/example.rs
