+++ mancala-tui/src/game.rs
@@ -1,0 +1,14 @@
+use mancala;
+
+use cursive::{views::Dialog, Cursive};
+
+pub fn run(){
+    // let mut siv = cursive::default();
+    
+    // siv.add_global_callback('q', Cursive::quit);
+
+    // siv.add_layer(Dialog::text("Hey!"));
+    
+    // siv.run();
+    Mancala
+}
