--- mancala/src/game.rs
@@ -1,11 +1,0 @@
-use cursive::{views::Dialog, Cursive};
-
-pub fn run(){
-    let mut siv = cursive::default();
-    
-    siv.add_global_callback('q', Cursive::quit);
-
-    siv.add_layer(Dialog::text("Hey!"));
-    
-    siv.run();
-}
